move to mocks directory to resolve jests manual-mock convention-s
